infra/vscode_web/requirements.txt
azure-ai-projects==1.0.0
azure-identity==1.19.0
openai
